Connection logic moved to MainPage along with syncing logc. Views not completely in sync for writing to keyboard.
