+++ AnneProKeyboard/LightingPage.xaml
@@ -161,7 +161,7 @@
             </Grid>
             <StackPanel Orientation="Horizontal">
                 <Grid>
-                    <Button x:Name="LightSyncButton" Content="Sync Profile" HorizontalAlignment="Right" VerticalAlignment="Bottom" Margin="0, 0, 10, 0" IsEnabled="False" Click="KeyboardSyncButton_Click" Height="32" Width="100" />
+                    <Button x:FieldModifier="public" x:Name="LightSyncButton" Content="Sync Profile" HorizontalAlignment="Right" VerticalAlignment="Bottom" Margin="0, 0, 10, 0" IsEnabled="False" Click="KeyboardSyncButton_Click" Height="32" Width="100" />
                     <StackPanel Orientation="Vertical" Margin="0,0,0,37">
                         <StackPanel Orientation="Horizontal" Margin="0,8,0,0" Grid.Row="3">
                             <Button Name="AllKeys" Content="All Keys" VerticalAlignment="Bottom" Margin="10, 0, 0, 0" IsEnabled="True" Click="KeyboardAllButton_Click" Height="32" Width="100" />
